Добавил новый класс и интерфейс
