+++ src/manager/InMemoryTaskManager.java
@@ -1,0 +1,172 @@
+package manager;
+
+import model.Epic;
+import model.Status;
+import model.Subtask;
+import model.Task;
+
+import java.util.*;
+
+public class InMemoryTaskManager implements TaskManager {
+    private final Map<Integer, Task> tasks = new HashMap<>();
+    private final Map<Integer, Epic> epics = new HashMap<>();
+    private final Map<Integer, Subtask> subtasks = new HashMap<>();
+    private int currentId = 1;
+
+    private int generateId() {
+        return currentId++;
+    }
+
+    public List<Task> getAllTasks() {
+        return new ArrayList<>(tasks.values());
+    }
+
+    public List<Epic> getAllEpics() {
+        return new ArrayList<>(epics.values());
+    }
+
+    public List<Subtask> getAllSubtasks() {
+        return new ArrayList<>(subtasks.values());
+    }
+
+    public void clearAllTasks() {
+        tasks.clear();
+    }
+
+    public void clearAllEpics() {
+        epics.clear();
+        subtasks.clear();
+    }
+
+    public void clearAllSubtasks() {
+        for (Epic epic : epics.values()) {
+            epic.getSubtaskIds().clear();
+        }
+        subtasks.clear();
+    }
+
+    public Task getTaskById(int id) {
+        return tasks.get(id);
+    }
+
+    public Epic getEpicById(int id) {
+        return epics.get(id);
+    }
+
+    public Subtask getSubtaskById(int id) {
+        return subtasks.get(id);
+    }
+
+    public int addTask(Task task) {
+        int id = generateId();
+        task.setId(id);
+        tasks.put(id, task);
+        return id;
+    }
+
+    public int addEpic(Epic epic) {
+        int id = generateId();
+        epic.setId(id);
+        epics.put(id, epic);
+        return id;
+    }
+
+    public int addSubtask(Subtask subtask) {
+        int epicId = subtask.getEpicId();
+        Epic epic = epics.get(epicId);
+        if (epic == null) {
+            return -1;
+        }
+        int id = generateId();
+        subtask.setId(id);
+        subtasks.put(id, subtask);
+        epic.addSubtaskId(id);
+        updateEpicStatus(epic);
+        return id;
+    }
+
+    public void updateTask(Task task) {
+        if (tasks.containsKey(task.getId())) {
+            tasks.put(task.getId(), task);
+        }
+    }
+
+    public void updateEpic(Epic epic) {
+        Epic oldEpic = epics.get(epic.getId());
+        if (oldEpic != null) {
+            oldEpic.setName(epic.getName());
+            oldEpic.setDescription(epic.getDescription());
+        }
+    }
+
+    public void updateSubtask(Subtask subtask) {
+        if (subtasks.containsKey(subtask.getId())) {
+            subtasks.put(subtask.getId(), subtask);
+            Epic epic = epics.get(subtask.getEpicId());
+            if (epic != null) {
+                updateEpicStatus(epic);
+            }
+        }
+    }
+
+    public void deleteTaskById(int id) {
+        tasks.remove(id);
+    }
+
+    public void deleteEpicById(int id) {
+        Epic epic = epics.remove(id);
+        if (epic != null) {
+            for (Integer subId : epic.getSubtaskIds()) {
+                subtasks.remove(subId);
+            }
+        }
+    }
+
+    public void deleteSubtaskById(int id) {
+        Subtask subtask = subtasks.remove(id);
+        if (subtask != null) {
+            Epic epic = epics.get(subtask.getEpicId());
+            if (epic != null) {
+                epic.removeSubtaskId(id);
+                updateEpicStatus(epic);
+            }
+        }
+    }
+
+    public List<Subtask> getSubtasksOfEpic(int epicId) {
+        List<Subtask> result = new ArrayList<>();
+        Epic epic = epics.get(epicId);
+        if (epic != null) {
+            for (Integer subId : epic.getSubtaskIds()) {
+                Subtask sub = subtasks.get(subId);
+                if (sub != null) {
+                    result.add(sub);
+                }
+            }
+        }
+        return result;
+    }
+
+    private void updateEpicStatus(Epic epic) {
+        List<Integer> subIds = epic.getSubtaskIds();
+        if (subIds.isEmpty()) {
+            epic.setStatus(Status.NEW);
+            return;
+        }
+        boolean allNew = true;
+        boolean allDone = true;
+        for (Integer id : subIds) {
+            Status status = subtasks.get(id).getStatus();
+            if (status != Status.NEW) allNew = false;
+            if (status != Status.DONE) allDone = false;
+        }
+        if (allNew) {
+            epic.setStatus(Status.NEW);
+        } else if (allDone) {
+            epic.setStatus(Status.DONE);
+        } else {
+            epic.setStatus(Status.IN_PROGRESS);
+        }
+    }
+}
+
